feat: move the `pass_through` property up to the parent `Window` class, rendering all subclasses with the same property
