+++ fabric/widgets/x11.py
@@ -156,6 +156,7 @@ class X11Window(Window):
         title: str = "fabric",
         type: Literal["top-level", "popup"] | Gtk.WindowType = Gtk.WindowType.TOPLEVEL,
         child: Gtk.Widget | None = None,
+        pass_through: bool = False,
         name: str | None = None,
         visible: bool = True,
         all_visible: bool = False,
@@ -179,6 +180,7 @@ class X11Window(Window):
             title,
             type,
             child,
+            pass_through,
             name,
             False,
             False,
